[FIX] Moved data models to separate folders
