--- backend/dev/dataModel/test/testModel.ts
@@ -1,12 +1,0 @@
-import { prop, Typegoose } from '@hasezoey/typegoose';
-
-export default class TestModel extends Typegoose {
-
-    @prop({ required: true })
-    test: string;
-
-    constructor(test:string = '') {
-        super();
-        this.test = test;
-    }
-}
