+++ backend/dev/server/repositories/movieRepositories/MovieRepository.ts
@@ -1,4 +1,4 @@
-import Movie from "../../../dataModel/test/movies/Movie";
+import Movie from "../../../dataModel/movies/Movie";
 
 export const MovieRepository = new Movie().getModelForClass(Movie);
